update test branch

+++ libmx/src/mx_strncmp.c
@@ -2,9 +2,9 @@
 
 int mx_strncmp(const char *src, const char *str, size_t n)
 {
-    size_t count = 0;
+    long long count = 0;
 
-    for(; count < (n - 1) && !(src[count] - str[count]); ++count);
+    for(; count < (long long)(n - 1) && !(src[count] - str[count]); ++count);
 
     return (src[count] - str[count]);
 }
